algorithm-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1

torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.206
transformers==4.35.2
accelerate==0.25.0
Pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
qwen-vl-utils==0.0.1

loguru==0.7.2
